refactor(solver): add optimal move strategy to Node struct
Add the `OptimalMove` strategy to the `Node` struct in the `solver.rs` file. This allows for finding the move optimal solution in addition to the push optimal solution. The `Node` struct now includes a `moves` field to track the number of moves made. This change enhances the flexibility and functionality of the solver module.